+++ src/main/java/com/cloudera/utils/hms/mirror/domain/StorageMigration.java
@@ -53,7 +53,7 @@ public class StorageMigration implements Cloneable {
     private boolean consolidateTablesForDistcp = Boolean.FALSE;
     @Schema(description = "When strict is true, any issues during evaluation will cause the migration to fail. When false, " +
             "the migration will continue but the issues will be reported. This can lead to data movement issues.")
-    private boolean strict = Boolean.TRUE;
+    private boolean strict = Boolean.FALSE;
 
     @Override
     public StorageMigration clone() {