+++ src/test/java/com/cloudera/utils/hms/mirror/integration/end_to_end/cdp/Test_sm_wd_epl_dc_strict.java
@@ -17,7 +17,6 @@
 
 package com.cloudera.utils.hms.mirror.integration.end_to_end.cdp;
 
-
 import com.cloudera.utils.hms.mirror.PhaseState;
 import com.cloudera.utils.hms.mirror.cli.Mirror;
 import com.cloudera.utils.hms.mirror.domain.support.Environment;
@@ -34,44 +33,33 @@ import static org.junit.Assert.assertEquals;
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = Mirror.class,
         args = {
+                "--hms-mirror.config.target-namespace=ofs://OHOME90",
                 "--hms-mirror.conversion.test-filename=/test_data/ext_purge_odd_parts_01.yaml",
-                "--hms-mirror.config.filename=/config/default.yaml.cdp",
-                "--hms-mirror.config.storage-migration-strict=false", // This would be 'false' by default.
-                "--hms-mirror.config.output-dir=${user.home}/.hms-mirror/test-output/e2e/cdp/sm_smn_wd_epl_dc_mismatch",
-        })
-@ActiveProfiles("e2e-cdp-sm_smn_wd_epl_dc")
+                "--hms-mirror.config.storage-migration-strict=true",
+                "--hms-mirror.config.output-dir=${user.home}/.hms-mirror/test-output/e2e/cdp/sm_wd_epl_dc_strict"
+        }
+)
+@ActiveProfiles("e2e-cdp-sm_wd_epl_dc")
 @Slf4j
-/*
-This test uses a configuration (LEFT hcfs namespace) that doesn't match the table/partition prefixes. This should
-fail with warnings.
-
-FIXED:1. when namespace in table doesn't match the namespace specified in the hcfsNamespace, nothing is translated.
-    - This should result in an error and warnings about why this didn't work.
-
-TODO: We need to fix the return code to be negative on Errors and Positive on 'table' conversion failures
-        but success app run.
- */
-public class Test_sm_smn_wd_epl_dc_mismatch extends E2EBaseTest {
-
+public class Test_sm_wd_epl_dc_strict extends E2EBaseTest {
     /*
             String[] args = new String[]{"-d", "STORAGE_MIGRATION",
                 "-wd", "/finance/managed-fso",
                 "-ewd", "/finance/external-fso",
-                "-smn", "ofs://OHOME90",
+//                "-smn", "ofs://OHOME90",
                 "-epl",
+//                "-rdl",
                 "-dc",
-                "-ltd", EXT_PURGE_ODD_PARTS_03, "-cfg", CDP,
+                "-ltd", EXT_PURGE_ODD_PARTS_03, "-cfg", CDP_CDP,
                 "-o", outputDir
         };
         long rtn = 0;
         MirrorLegacy mirror = new MirrorLegacy();
         rtn = mirror.go(args);
-        assertEquals("Return Code Failure: " + rtn, 1, rtn);
+        long check = STORAGE_MIGRATION_NAMESPACE_LEFT_MISSING_RDL_GLM.getLong();
 
-        // Read the output and verify the results.
-        DBMirror[] resultsMirrors = getResults(outputDir, EXT_PURGE_ODD_PARTS_03);
+        assertEquals("Return Code Failure: " + rtn + " doesn't match: " + check * -1, check * -1, rtn);
 
-        validatePhase(resultsMirrors[0], "web_sales", PhaseState.ERROR);
 
      */
     @Test
@@ -79,13 +67,13 @@ public class Test_sm_smn_wd_epl_dc_mismatch extends E2EBaseTest {
         // Get Runtime Return Code.
         long rtn = getReturnCode();
         // Verify the return code.
-        long check = 0L;
+        long check = 1L;
         assertEquals("Return Code Failure: " + rtn, check, rtn);
     }
 
     @Test
     public void phaseTest() {
-        validatePhase("ext_purge_odd_parts", "web_sales", PhaseState.SUCCESS);
+        validatePhase("ext_purge_odd_parts", "web_sales", PhaseState.ERROR);
     }
 
     @Test
@@ -93,4 +81,5 @@ public class Test_sm_smn_wd_epl_dc_mismatch extends E2EBaseTest {
         validateTableIssueCount("ext_purge_odd_parts", "web_sales",
                 Environment.LEFT, 2);
     }
+
 }